+++ hpc/nvidia/cuda/file/gdsio_awk.sh
@@ -1,0 +1,23 @@
+#!/bin/bash
+set -e
+
+cat gdsio.log | awk '{
+		for (i=1; i<=NF; i++) {
+			if ($i == "IoType:") {
+				gsub(/,/, "", $(i+1));
+				iotype=$(i+1);
+			}
+			if ($i == "Threads:") {
+				# remove ','
+				gsub(/,/, "", $(i+1));
+				threads=$(i+1)
+			}
+			if ($i == "IOSize:") {
+				iosize=$(i+1);
+			}
+			if ($i == "Throughput:") {
+				throughput=$(i+1);
+			}
+		}
+		printf("%s %s %s\n", iosize / 1024, threads, throughput);
+	}'